Improve code and documentation
